+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/BanyanDBConfigLoader.java
@@ -20,12 +20,15 @@ package org.apache.skywalking.oap.server.storage.plugin.banyandb;
 
 import java.io.FileNotFoundException;
 import java.io.Reader;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.server.library.module.ModuleProvider;
 import org.apache.skywalking.oap.server.library.module.ModuleStartException;
+import org.apache.skywalking.oap.server.library.util.CollectionUtils;
 import org.apache.skywalking.oap.server.library.util.ResourceUtils;
 import org.yaml.snakeyaml.Yaml;
 
@@ -57,26 +60,8 @@ public class BanyanDBConfigLoader {
 
         Map<String, Properties> configProperties = new HashMap<>();
         configMap.forEach((part, c) -> {
-            final Properties properties = new Properties();
             if (c != null) {
-                for (Map.Entry<String, ?> entry : c.entrySet()) {
-                    String propertyName = entry.getKey();
-                    Object propertyValue = entry.getValue();
-                    if (propertyValue instanceof Map) {
-                        Properties subProperties = new Properties();
-                        for (Map.Entry<String, ?> e : ((Map<String, ?>) propertyValue).entrySet()) {
-                            String key = e.getKey();
-                            Object value = e.getValue();
-                            subProperties.put(key, value);
-                            replacePropertyAndLog(key, value, subProperties, this.moduleProvider.name(), yaml);
-                        }
-                        properties.put(propertyName, subProperties);
-                    } else {
-                        properties.put(propertyName, propertyValue);
-                        replacePropertyAndLog(
-                            propertyName, propertyValue, properties, this.moduleProvider.name(), yaml);
-                    }
-                }
+                final Properties properties = parseConfig(c);
                 configProperties.put(part, properties);
             }
         });
@@ -87,26 +72,62 @@ public class BanyanDBConfigLoader {
                 moduleProvider.name()
             );
             Properties groups = configProperties.get("groups");
+            Properties recordsNormal = (Properties) groups.get("recordsNormal");
             copyProperties(
-                config.getRecordsNormal(), (Properties) groups.get("recordsNormal"),
+                config.getRecordsNormal(), recordsNormal,
                 moduleProvider.getModule().name(), moduleProvider.name()
             );
+            config.getRecordsNormal()
+                  .setLifecycleStages(copyStages(
+                      "recordsNormal", recordsNormal, config.getRecordsNormal().getEnabledAdditionalStages(),
+                      config.getRecordsNormal().getDefaultQueryStages()
+                  ));
+
+            Properties recordsSupper = (Properties) groups.get("recordsSuper");
             copyProperties(
-                config.getRecordsSuper(), (Properties) groups.get("recordsSuper"),
+                config.getRecordsSuper(), recordsSupper,
                 moduleProvider.getModule().name(), moduleProvider.name()
             );
+            config.getRecordsSuper()
+                  .setLifecycleStages(
+                      copyStages(
+                          "recordsSuper", recordsSupper, config.getRecordsSuper().getEnabledAdditionalStages(),
+                          config.getRecordsSuper().getDefaultQueryStages()
+                      ));
+
+            Properties metricsMin = (Properties) groups.get("metricsMin");
             copyProperties(
-                config.getMetricsMin(), (Properties) groups.get("metricsMin"),
+                config.getMetricsMin(), metricsMin,
                 moduleProvider.getModule().name(), moduleProvider.name()
             );
+            config.getMetricsMin()
+                  .setLifecycleStages(copyStages(
+                      "metricsMin", metricsMin, config.getMetricsMin().getEnabledAdditionalStages(),
+                      config.getMetricsMin().getDefaultQueryStages()
+                  ));
+
+            Properties metricsHour = (Properties) groups.get("metricsHour");
             copyProperties(
-                config.getMetricsHour(), (Properties) groups.get("metricsHour"),
+                config.getMetricsHour(), metricsHour,
                 moduleProvider.getModule().name(), moduleProvider.name()
             );
+            config.getMetricsHour()
+                  .setLifecycleStages(copyStages(
+                      "metricsHour", metricsHour, config.getMetricsHour().getEnabledAdditionalStages(),
+                      config.getMetricsHour().getDefaultQueryStages()
+                  ));
+
+            Properties metricsDay = (Properties) groups.get("metricsDay");
             copyProperties(
-                config.getMetricsDay(), (Properties) groups.get("metricsDay"),
+                config.getMetricsDay(), metricsDay,
                 moduleProvider.getModule().name(), moduleProvider.name()
             );
+            config.getMetricsDay()
+                  .setLifecycleStages(copyStages(
+                      "metricsDay", metricsDay, config.getMetricsDay().getEnabledAdditionalStages(),
+                      config.getMetricsDay().getDefaultQueryStages()
+                  ));
+
             copyProperties(
                 config.getMetadata(), (Properties) groups.get("metadata"),
                 moduleProvider.getModule().name(), moduleProvider.name()
@@ -119,5 +140,62 @@ public class BanyanDBConfigLoader {
             throw new ModuleStartException("Failed to load BanyanDB configuration.", e);
         }
         return config;
+    }
+
+    private Properties parseConfig(final Map<String, ?> config) {
+        final Properties properties = new Properties();
+        for (Map.Entry<String, ?> entry : config.entrySet()) {
+            String propertyName = entry.getKey();
+            Object propertyValue = entry.getValue();
+            if (propertyValue instanceof Map) {
+                Properties subProperties = parseConfig((Map<String, ?>) propertyValue);
+                properties.put(propertyName, subProperties);
+            } else {
+                properties.put(propertyName, propertyValue);
+                replacePropertyAndLog(
+                    propertyName, propertyValue, properties, this.moduleProvider.name(), yaml);
+            }
+        }
+        return properties;
+    }
+
+    private List<BanyanDBStorageConfig.Stage> copyStages(final String configName,
+                                                         final Properties group,
+                                                         final List<String> stagesNames,
+                                                         final List<String> defaultQueryStages) throws IllegalAccessException {
+        if (CollectionUtils.isNotEmpty(defaultQueryStages)) {
+            for (String stageName : defaultQueryStages) {
+                if (BanyanDBStorageConfig.StageName.hot.name().equals(stageName)) {
+                    continue;
+                }
+                if (CollectionUtils.isEmpty(stagesNames) || !stagesNames.contains(stageName)) {
+                    throw new IllegalArgumentException("Group configuration: " + configName + ": " + group.toString() + " [defaultQueryStages] error, the stages [" + stageName + "] is not in [enabledStages].");
+                }
+            }
+        }
+
+        if (CollectionUtils.isNotEmpty(stagesNames)) {
+           List<BanyanDBStorageConfig.Stage> stages = new ArrayList<>(stagesNames.size());
+            if (stagesNames.contains(BanyanDBStorageConfig.StageName.warm.name())) {
+                BanyanDBStorageConfig.Stage warm = new BanyanDBStorageConfig.Stage();
+                warm.setName(BanyanDBStorageConfig.StageName.warm);
+                copyProperties(
+                    warm, (Properties) group.get(BanyanDBStorageConfig.StageName.warm.name()),
+                    moduleProvider.getModule().name(), moduleProvider.name()
+                );
+                stages.add(warm);
+            } else if (stagesNames.contains(BanyanDBStorageConfig.StageName.cold.name())) {
+                BanyanDBStorageConfig.Stage cold = new BanyanDBStorageConfig.Stage();
+                cold.setName(BanyanDBStorageConfig.StageName.cold);
+                copyProperties(
+                    cold, (Properties) group.get(BanyanDBStorageConfig.StageName.cold.name()),
+                    moduleProvider.getModule().name(), moduleProvider.name()
+                );
+                stages.add(cold);
+            }
+           return stages;
+       } else {
+           return List.of();
+       }
     }
 }